Split AsymptoticBondiData up into separate files
