+++ scri/asymptotic_bondi_data/__init__.py
@@ -5,21 +5,71 @@ from .. import ModesTimeSeries
 
 
 class AsymptoticBondiData(object):
-    def __init__(self, time, ell_max, multiplication_truncator=max):
+    """Class to store asymptotic Bondi data
+
+    This class stores time data, along with the corresponding values of psi0 through psi4 and sigma.
+    For simplicity, the data are stored as one contiguous array.  That is, *all* values are stored
+    at all times, even if they are zero, and all Modes objects are stored with ell_min=0, even when
+    their spins are not zero.
+
+    The single contiguous array is then viewed as 6 separate ModesTimeSeries objects, which enables
+    them to track their spin weights, and provides various convenient methods like `eth` and
+    `ethbar`; `dot` and `ddot` for time-derivatives; `int` and `iint` for time-integrations; `norm`
+    to take the norm of a function over the sphere; `bar` for conjugation of the functions (which is
+    different from just conjugating the mode weights); etc.  It also handles algebra correctly --
+    particularly addition (which is disallowed when the spin weights differ) and multiplication
+    (which can be delicate with regards to the resulting ell values).
+
+    This may lead to some headaches when the user tries to do things that are disabled by Modes
+    objects.  The goal is to create headaches if and only if the user is trying to do things that
+    really should never be done (like conjugating mode weights, rather than the underlying function;
+    adding modes with different spin weights; etc.).  Please open issues for any situations that
+    don't meet this standard.
+
+    This class also provides various convenience methods for computing things like the mass aspect,
+    the Bondi four-momentum, the Bianchi identities, etc.
+
+    """
+    def __init__(self, time, ell_max, multiplication_truncator=sum):
+        """Create new storage for asymptotic Bondi data
+
+        Parameters
+        ==========
+        time: int or array_like
+            Times at which the data will be stored.  If this is an int, an empty array of that size
+            will be created.  Otherwise, this must be a 1-dimensional array of floats.
+        ell_max: int
+            Maximum ell value to be stored
+        multiplication_truncator: callable [defaults to `sum`, even though `max` is nicer]
+            Function to be used by default when multiplying Modes objects together.  See the
+            documentation for spherical_functions.Modes.multiply for more details.  The default
+            behavior with `sum` is the most correct one -- keeping all ell values that result -- but
+            also the most wasteful, and very likely to be overkill.  The user should probably always
+            use `max`.  (Unfortunately, this must remain an opt-in choice, to ensure that the user
+            is aware of the situation.)
+
+        """
+        if np.ndim(time) == 0:
+            # Assume this is just the size of the time array; construct an empty array
+            time = np.empty((time,), dtype=float)
+        elif np.ndim(time) > 1:
+            raise ValueError(f"Input `time` parameter must be an integer or a 1-d array; it has shape {time.shape}")
+        if time.dtype != float:
+            raise ValueError(f"Input `time` parameter must have dtype float; it has dtype {time.dtype}")
         shape = [6, time.size, LM_total_size(0, ell_max)]
-        data = np.zeros(shape, dtype=complex)
         self._time = time.copy()
-        self._psi0 = ModesTimeSeries(data[0], self._time, spin_weight=2,
+        self._raw_data = np.zeros(shape, dtype=complex)
+        self._psi0 = ModesTimeSeries(self._raw_data[0], self._time, spin_weight=2,
                                      ell_max=ell_max, multiplication_truncator=multiplication_truncator)
-        self._psi1 = ModesTimeSeries(data[1], self._time, spin_weight=1,
+        self._psi1 = ModesTimeSeries(self._raw_data[1], self._time, spin_weight=1,
                                      ell_max=ell_max, multiplication_truncator=multiplication_truncator)
-        self._psi2 = ModesTimeSeries(data[2], self._time, spin_weight=0,
+        self._psi2 = ModesTimeSeries(self._raw_data[2], self._time, spin_weight=0,
                                      ell_max=ell_max, multiplication_truncator=multiplication_truncator)
-        self._psi3 = ModesTimeSeries(data[3], self._time, spin_weight=-1,
+        self._psi3 = ModesTimeSeries(self._raw_data[3], self._time, spin_weight=-1,
                                      ell_max=ell_max, multiplication_truncator=multiplication_truncator)
-        self._psi4 = ModesTimeSeries(data[4], self._time, spin_weight=-2,
+        self._psi4 = ModesTimeSeries(self._raw_data[4], self._time, spin_weight=-2,
                                      ell_max=ell_max, multiplication_truncator=multiplication_truncator)
-        self._sigma = ModesTimeSeries(data[5], self._time, spin_weight=2,
+        self._sigma = ModesTimeSeries(self._raw_data[5], self._time, spin_weight=2,
                                       ell_max=ell_max, multiplication_truncator=multiplication_truncator)
 
     @property
@@ -93,78 +143,33 @@ class AsymptoticBondiData(object):
         self._psi0[:] = psi0prm
         return self.psi0
 
-    @classmethod
-    def from_initial_values(cls, time, ell_max=8, sigma=0.0, sigmadot=0.0, sigmaddot=0.0, psi2=0.0, psi1=0.0, psi0=0.0):
-        """Construct Bondi data from sigma as a function of time and optional initial values
+    from .from_initial_values import from_initial_values
 
-        The initial-value formulation for Bondi gauge is determined by these relations:
+    def mass_aspect(self, truncate_ell=max):
+        """Compute the Bondi mass aspect of the AsymptoticBondiData
 
-            \dot{\Psi_0} = -\eth\Psi_1 + 3\sigma \Psi_2
-            \dot{\Psi_1} = -\eth\Psi_2 + 2\sigma \Psi_3
-            \dot{\Psi_2} = -\eth\Psi_3 + \sigma \Psi_4
-            \Psi_3 = \eth \dot{\bar{\sigma}}
-            \Psi_4 = - \ddot{\bar{\sigma}}
+        The Bondi mass aspect is given by
 
-        We also have a constraint on the initial value of Psi2:
+            \Psi = \psi_2 + \eth \eth \bar{\sigma} + \sigma * \dot{\bar{\sigma}}
 
-            Im[\Psi_2] = -Im[\eth^2\bar{\sigma} + \sigma \dot{\bar{\sigma}}]
+        Note that the last term is a product between two fields.  If, for example, these both have
+        ell_max=8, then their full product would have ell_max=16, meaning that we would go from
+        tracking 77 modes to 289.  This shows that deciding how to truncate the output ell is
+        important, which is why this function has the extra argument that it does.
+
+        Parameters
+        ==========
+        truncate_ell: int, or callable [defaults to `max`]
+            Determines how the ell_max value of the output is determined.  If an integer is passed,
+            each term in the output is truncated to have at most that ell_max.  (In particular,
+            terms that will not be used in the output are simply not computed, without incurring any
+            errors due to aliasing.)  If a callable is passed, it is passed on to the
+            spherical_functions.Modes.multiply method.  See that function's docstring for details.
+            The default behavior will result in the output having ell_max equal to the largest of
+            any of the individual Modes objects in the equation for \Psi above -- but not the
+            product.
 
         """
-
-        def asany_atleast2d_complex(a):
-            a = np.asanyarray(a) + 0j
-            while np.ndim(a) < 2:
-                a = a[np.newaxis, ...]
-            return a
-
-        psi2 = asany_atleast2d_complex(psi2)
-        psi1 = asany_atleast2d_complex(psi1)
-        psi0 = asany_atleast2d_complex(psi0)
-
-        # Construct the empty container
-        abd = AsymptoticBondiData(time, ell_max)
-
-        # Evaluate sigma and derivatives
-        if np.ndim(sigma) == 0 or np.ndim(sigma) == 1:
-            # Assume this is just the angular dependence, which will be taken as constant in time.
-            # If this is true, assumes sigmadot and sigmaddot are constants in time, and just
-            # integrates them.
-            # sigma = asany_atleast2d_complex(sigma)
-            # sigmadot = asany_atleast2d_complex(sigmadot)
-            # sigmaddot = asany_atleast2d_complex(sigmaddot)
-            # abd.sigma = sigma
-            # abd.sigma = abd.sigma + abd.time * (sigmadot + abd.time * (sigmaddot / 2))
-            sigma = ModesTimeSeries(sigma+0j, abd.time, spin_weight=2)
-            sigmadot = ModesTimeSeries(sigmadot+0j, abd.time, spin_weight=2)
-            sigmaddot = ModesTimeSeries(sigmaddot+0j, abd.time, spin_weight=2)
-            abd.sigma = sigma + abd.time * (sigmadot + abd.time * (sigmaddot / 2))
-        elif np.ndim(sigma) == 2:
-            # Assume this gives complete data, as a function of time and angle.
-            # If this is true, ignore sigmadot and sigmaddot.
-            abd.sigma = sigma
-            sigmadot = abd.sigma.dot
-            sigmaddot = abd.sigma.ddot
-        else:
-            raise ValueError(f"Input `sigma` must have 1 or 2 dimensions; it has {np.ndim(sigma)}")
-
-        # Adjust the initial value of psi2 to satisfy the mass-aspect condition
-        sigma_initial = abd.sigma[..., 0, :]
-        sigma_bar_dot_initial = abd.sigma.bar.dot[..., 0, :]
-        psi2 = (
-            ModesTimeSeries(psi2, abd.time, spin_weight=0).real
-            - (sigma_initial.bar.eth.eth + sigma_initial * sigma_bar_dot_initial).imag
-        )
-
-        # Compute the Weyl components
-        abd.psi4 = -sigmaddot.bar
-        abd.psi3 = sigmadot.bar.eth
-        abd.psi2 = (-abd.psi3.eth +     abd.sigma * abd.psi4).int + psi2
-        abd.psi1 = (-abd.psi2.eth + 2 * abd.sigma * abd.psi3).int + psi1
-        abd.psi0 = (-abd.psi1.eth + 3 * abd.sigma * abd.psi2).int + psi0
-
-        return abd
-
-    def mass_aspect(self, truncate_ell=None):
         if callable(truncate_ell):
             return self.psi2 + self.sigma.bar.eth.eth + self.sigma.multiply(self.sigma.bar.dot, truncator=truncate_ell)
         elif truncate_ell:
@@ -178,254 +183,15 @@ class AsymptoticBondiData(object):
 
     @property
     def bondi_four_momentum(self):
+        """Compute the Bondi four-momentum of the AsymptoticBondiData"""
         Psi_restricted = self.mass_aspect(1).view(np.ndarray).real  # Compute only the parts of the mass aspect we need, ell<=1
         four_momentum = np.empty(Psi_restricted.shape, dtype=float)
         four_momentum[..., 0] = - Psi_restricted[..., 0] / math.sqrt(8)
         four_momentum[..., 1:4] = - Psi_restricted[..., 1:4] / 6
         return four_momentum
 
-    def bondi_constraints(self, lhs=True, rhs=True):
-        """Compute Bondi-gauge constraint equations
-
-        Bondi gauge establishes some relations that the data must satisfy:
-
-            \dot{\Psi0} = -\eth\Psi_1 + 3\sigma \Psi_2
-            \dot{\Psi1} = -\eth\Psi_2 + 2\sigma \Psi_3
-            \dot{\Psi2} = -\eth\Psi_3 + \sigma \Psi_4
-            \Psi_3 = \eth \dot{\bar{\sigma}}
-            \Psi_4 = - \ddot{\bar{\sigma}}
-            Im[\Psi_2] = -Im[\eth^2\bar{\sigma} + \sigma \dot{\bar{\sigma}}]
-
-        This function returns a 6-tuple of 2-tuples, corresponding to these 6 equations and their
-        left- and right-hand sides.
-
-        """
-        return (
-            self.bianchi_0(lhs, rhs),
-            self.bianchi_1(lhs, rhs),
-            self.bianchi_2(lhs, rhs),
-            self.constraint_3(lhs, rhs),
-            self.constraint_4(lhs, rhs),
-            self.constraint_mass_aspect(lhs, rhs),
-        )
-
-    @property
-    def bondi_violations(self):
-        """Compute violations of Bondi-gauge constraints
-
-        Bondi gauge establishes some relations that the data must satisfy:
-
-            \dot{\Psi0} = -\eth\Psi_1 + 3\sigma \Psi_2
-            \dot{\Psi1} = -\eth\Psi_2 + 2\sigma \Psi_3
-            \dot{\Psi2} = -\eth\Psi_3 + \sigma \Psi_4
-            \Psi_3 = \eth \dot{\bar{\sigma}}
-            \Psi_4 = - \ddot{\bar{\sigma}}
-            Im[\Psi_2] = -Im[\eth^2\bar{\sigma} + \sigma \dot{\bar{\sigma}}]
-
-        This function returns a tuple of 6 arrays, corresponding to these 6 equations, in which the
-        right-hand side is subtracted from the left-hand side.  No norms are taken.
-
-        """
-        constraints = self.bondi_constraints(True, True)
-        return (lhs-rhs for (lhs, rhs) in constraints)
-
-    @property
-    def bondi_violation_norms(self):
-        """Compute norms of violations of Bondi-gauge conditions
-
-        Bondi gauge establishes some relations that the data must satisfy:
-
-            \dot{\Psi0} = -\eth\Psi_1 + 3\sigma \Psi_2
-            \dot{\Psi1} = -\eth\Psi_2 + 2\sigma \Psi_3
-            \dot{\Psi2} = -\eth\Psi_3 + \sigma \Psi_4
-            \Psi_3 = \eth \dot{\bar{\sigma}}
-            \Psi_4 = - \ddot{\bar{\sigma}}
-            Im[\Psi_2] = -Im[\eth^2\bar{\sigma} + \sigma \dot{\bar{\sigma}}]
-
-        This function returns a tuple of 6 arrays, corresponding to the norms of these 6 equations,
-        in which the right-hand side is subtracted from the left-hand side, and then the squared
-        magnitude of that result is integrated over the sphere.  No integration is performed over
-        time.
-
-        """
-        violations = self.bondi_violations
-        return (v.norm() for v in violations)
-
-    def bianchi_0(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the Psi0 component of the Bianchi identity
-
-        In Bondi coordinates, the Bianchi identities simplify, resulting in this expression (among
-        others) for the time derivative of Psi0:
-
-            \dot{\Psi0} = -\eth\Psi_1 + 3\sigma \Psi_2
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = self.psi0.dot
-        if rhs:
-            rhs_value = -self.psi1.eth + 3 * self.sigma * self.psi2
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
-
-    def bianchi_1(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the Psi1 component of the Bianchi identity
-
-        In Bondi coordinates, the Bianchi identities simplify, resulting in this expression (among
-        others) for the time derivative of Psi1:
-
-            \dot{\Psi1} = -\eth\Psi_2 + 2\sigma \Psi_3
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = self.psi1.dot
-        if rhs:
-            rhs_value = -self.psi2.eth + 2 * self.sigma * self.psi3
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
-
-    def bianchi_2(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the Psi2 component of the Bianchi identity
-
-        In Bondi coordinates, the Bianchi identities simplify, resulting in this expression (among
-        others) for the time derivative of Psi2:
-
-            \dot{\Psi2} = -\eth\Psi_3 + \sigma \Psi_4
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = self.psi2.dot
-        if rhs:
-            rhs_value = -self.psi3.eth + self.sigma * self.psi4
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
-
-    def constraint_3(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the Psi3 expression in Bondi gauge
-
-        In Bondi coordinates, the value of Psi3 is given by a time derivative and an angular
-        derivative of the (conjugate) shear:
-
-            \Psi3 = \eth \dot{\bar{\sigma}}
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = self.psi3
-        if rhs:
-            rhs_value = self.sigma.bar.dot.eth
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
-
-    def constraint_4(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the Psi4 expression in Bondi gauge
-
-        In Bondi coordinates, the value of Psi4 is given by two time derivatives of the (conjugate)
-        shear:
-
-            \Psi4 = - \ddot{\bar{\sigma}}
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = self.psi4
-        if rhs:
-            rhs_value = -self.sigma.bar.ddot
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
-
-    def constraint_mass_aspect(self, lhs=True, rhs=True):
-        """Return the left- and/or right-hand sides of the mass-aspect reality condition in Bondi gauge
-
-        In Bondi coordinates, the Bondi mass aspect is always real, resulting in this relationship:
-
-            \Im[\Psi2] = - \Im[\eth^2\bar{\sigma} + \sigma \dot{\bar{\sigma}}]
-
-        Parameters
-        ==========
-        lhs: bool [defaults to True]
-            If True, return the left-hand side of the equation above
-        rhs: bool [defaults to True]
-            If True, return the right-hand side of the equation above
-
-        If both parameters are True, a tuple with elements (lhs_value, rhs_value) is returned;
-        otherwise just the requested value is returned.
-
-        """
-        if lhs:
-            lhs_value = np.imag(self.psi2.view(np.ndarray))
-        if rhs:
-            rhs_value = -np.imag(self.sigma.bar.eth.eth + self.sigma * self.sigma.bar.dot)
-        if lhs and rhs:
-            return (lhs_value, rhs_value)
-        elif lhs:
-            return lhs_value
-        elif rhs:
-            return rhs_value
+    from .constraints import (
+        bondi_constraints, bondi_violations, bondi_violation_norms,
+        bianchi_0, bianchi_1, bianchi_2,
+        constraint_3, constraint_4, constraint_mass_aspect
+    )
